Convert test_hessian_normalise into unit test; rename test_riemannianmetric to replace test_metric
